CalculatorApp/Service/BasicArithmetics.cs
using Calculator.Interface;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Calculator.Service
{
    public class BasicArithmetics : IBasicArithmetics
    {
        public double Addition(double num1, double num2)
        {
            return num1 + num2;
        }

        public double Subtraction(double num1, double num2)
        {
            return num1 - num2;
        }

        public double Multiplication(double num1, double num2)
        {
            return num1 * num2;
        }

        public double Division(double num1, double num2)
        {
            if (num2 == 0) 
                throw new DivideByZeroException("Can't use 0 as denominator");

            return num1 / num2;
        }
    }
}
